test: replace CAREamistV2 tests with init from checkpoint test; mv fixtures to conftest
